# Optimize decompress generic