The current decompress_generic code suffers from many blocked loads, originally mentioned in issue #411. This patchset attemps to fix that issue.  Also similar issues mentioned in pull #470 , issue #126, and probably others.

The heart of the issue is that LZ4_wildCopy does a copy by 8 bytes at a time, leading to lots of work happening side-by-side.   The farther away we can move these copies by extending the copy width, the faster things seem to go.  Increasing the copy width to as wide as 32 bytes results in overall decompression speed improvement of around ~10% a xeon, ~20% on a modern AMD.

The primary code challenge then is that we have only 8 bytes of WILDCOPYLENGTH, when we'd really like far larger.  So we end up duplicating most of the decode loop, to add a fastpath for a larger copy length, and then the last handful of bytes can use the original loop.  

I've split up the code in what I hope are readable small diffs.  Testing has mostly consisted of 'make test', and an attempt at fuzzing with afl.  Has *not* been tested on other arches, but it would be fairly easy to #ifdef the first loop away and be the same as previous if it shows a regression.

Benchmarks on Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz, from lzbench.  The final percent is how much faster this patchset is vs. the current dev branch
```
./lzbench -elz4 silesia/*

lz4 1.8.3                 297 MB/s  2598 MB/s     6428742  63.07 silesia/dickens
lz4 1.8.3                 328 MB/s  2452 MB/s     6428742  63.07 silesia/dickens    -5% 

lz4 1.8.3                 472 MB/s  2374 MB/s    26435667  51.61 silesia/mozilla
lz4 1.8.3                 493 MB/s  2658 MB/s    26435667  51.61 silesia/mozilla        12%

lz4 1.8.3                 454 MB/s  2778 MB/s     5440937  54.57 silesia/mr
lz4 1.8.3                 493 MB/s  3037 MB/s     5440937  54.57 silesia/mr             9%

lz4 1.8.3                 796 MB/s  3240 MB/s     5533040  16.49 silesia/nci
lz4 1.8.3                 877 MB/s  4061 MB/s     5533040  16.49 silesia/nci            25%

lz4 1.8.3                 385 MB/s  2383 MB/s     4338918  70.53 silesia/ooffice
lz4 1.8.3                 426 MB/s  2709 MB/s     4338918  70.53 silesia/ooffice        14%

lz4 1.8.3                 413 MB/s  2463 MB/s     5256666  52.12 silesia/osdb
lz4 1.8.3                 439 MB/s  2950 MB/s     5256666  52.12 silesia/osdb           20%

lz4 1.8.3                 303 MB/s  2181 MB/s     3181387  48.00 silesia/reymont
lz4 1.8.3                 331 MB/s  2182 MB/s     3181387  48.00 silesia/reymont        0%

lz4 1.8.3                 511 MB/s  2806 MB/s     7716839  35.72 silesia/samba
lz4 1.8.3                 552 MB/s  3330 MB/s     7716839  35.72 silesia/samba          19%

lz4 1.8.3                 381 MB/s  2899 MB/s     6790273  93.63 silesia/sao
lz4 1.8.3                 412 MB/s  3586 MB/s     6790273  93.63 silesia/sao            24%

lz4 1.8.3                 448 MB/s  2475 MB/s   100881331  47.60 silesia/silesia.tar
lz4 1.8.3                 474 MB/s  2784 MB/s   100881331  47.60 silesia/silesia.tar    12%

lz4 1.8.3                 344 MB/s  2395 MB/s    20139988  48.58 silesia/webster
lz4 1.8.3                 372 MB/s  2536 MB/s    20139988  48.58 silesia/webster        6%

lz4 1.8.3                1104 MB/s  7769 MB/s     8390195  99.01 silesia/x-ray
lz4 1.8.3                1032 MB/s  9056 MB/s     8390195  99.01 silesia/x-ray          17%

lz4 1.8.3                 715 MB/s  2932 MB/s     1227495  22.96 silesia/xml
lz4 1.8.3                 729 MB/s  3514 MB/s     1227495  22.96 silesia/xml            20%
```

An all-zeroes test is particularly ugly for the current code:

```

dd if=/dev/zero of=zerotest bs=1M count=2
(Currently any match extending in to the last 64 bytes uses the slowpath, TODO)
dd if=/dev/random bs=65 count=1 >> zerotest

lz4 1.8.3               15996 MB/s  4920 MB/s        8296   0.40 silesia/zerotest
lz4 1.8.3               16041 MB/s 10151 MB/s        8296   0.40 silesia/zerotest       106%
```

I also tested on a AMD 2400G:

```
dickens          7%
mozilla          12%
mr               18%
nci              29%
ooffice          23%
osdb             28%
reymont          7%
samba            25%
sao              39%
silesia.tar      20%
webster          17%
xml              32%
x-ray            6%
```

The same issue likely exists in zstd (where it is likely easier to fix, since parsing isn't mixed up with sequence execution).