@Cyan4973 2019-01-29
Looks great @djwatson !

I noticed some non-trivial differences in compression speed, according to your measurements.
Are they related to your patch too, or is it measurement noise ?