@Cyan4973 2019-01-29
I confirm big decompression speed gains in benchmark measurements on `silesia`.
On my laptop, with `clang`, decompression speed went up, from 2800 MB/s to 3000 MB/s.
On my desktop, with `gcc`, gains were more dramatic, from 3150 MB/s to 3700 MB/s.
Desktop + gcc8 + `calgary` : from 3200 MB/s to 3550 MB/s.

However, in all tests, compression speed was completely stable.
I don't understand yet why your measurements seem to show some differences on compression side.