@djwatson 2019-01-29
Any compression speed changes are noise.