@Cyan4973 2019-01-30
Thanks for all these interesting data points @djwatson .
Indeed, the improvements related to overlapping copies are much larger than expectation. It's great !

One last question : 
the new decoder seems to completely remove the previous "2-stages shortcut", which is designed to special-case "common" small_run+small_matches.

Is that intentional ?

From a high level, it seems your optimizations, which target long copies, and overlapping copies, feel complementary with the shortcut, which specifically avoids these cases.