@Cyan4973 2019-02-01
> For match copy lengths that don't overflow, it did a memcopy by 8+8+2. We still do this, but 16+2.

There is a bit more to this story.
By doing `8+8+2`, this makes the shortcut compatible
with offsets in the [8-15] range.

In contrast, doing `16+2` requires the offset to be at least 16+.

The second version saves one load instruction per sequence,
but the first version makes more sequences compatible with the shortcut.
It might not seem much, but this triggers an impact on branch predictability.

The shorcut is expected to be useful "as often as possible".
Restraining conditions of compatible sequences slightly reduces its predictability,
so now there are a bit more branch misses, hence more pipeline flushes, etc.

With modern cpus, the cost of branch unpredictability is much higher that the cost of an instruction.
So it favors the `8+8+2` trade off.