@djwatson 2019-02-01
> I just made a test of the proposed PR on an ARM platform.
> Unfortunately, results are not so good on this platform :

Unfortunately I don't have any qualcomm arm chips, I'll have to borrow one to dig in to it. 

Alternatively, we could #ifdef this to only x86, but it looked like there were some nice wins on larger arm chips, it would be nice to fix it. 

> By doing 8+8+2, this makes the shortcut compatible with offsets in the [8-15] range.
> The second version saves one load instruction per sequence, but the first version makes more sequences compatible with the shortcut. It might not seem much, but this triggers an impact on branch predictability.

It's even slightly more complicated:  The first version (8+8+2) potentially triggers the load blocked by store condition, while doing it by-16 doesn't seem to.  However we don't do anything special to avoid this in memcpy_using_offset for 8-15 (currently only 1,2,4), although we could, but for the shortcut it's probably too small to worry about anyway. 

Benchmarks:

```
this branch:
lz4 1.8.3                 325 MB/s  3074 MB/s     6428742  63.07 silesia/dickens
lz4 1.8.3                 527 MB/s  2808 MB/s    26435667  51.61 silesia/mozilla
lz4 1.8.3                 515 MB/s  3522 MB/s     5440937  54.57 silesia/mr
lz4 1.8.3                 886 MB/s  4299 MB/s     5533040  16.49 silesia/nci
lz4 1.8.3                 436 MB/s  2885 MB/s     4338918  70.53 silesia/ooffice
lz4 1.8.3                 468 MB/s  3198 MB/s     5256666  52.12 silesia/osdb
lz4 1.8.3                 334 MB/s  2576 MB/s     3181387  48.00 silesia/reymont
lz4 1.8.3                 569 MB/s  3581 MB/s     7716839  35.72 silesia/samba
lz4 1.8.3                 447 MB/s  4133 MB/s     6790273  93.63 silesia/sao
lz4 1.8.3                 475 MB/s  3013 MB/s   100881331  47.60 silesia/silesia.tar
lz4 1.8.3                 379 MB/s  2857 MB/s    20139988  48.58 silesia/webster
lz4 1.8.3                1236 MB/s  9404 MB/s     8390195  99.01 silesia/x-ray
lz4 1.8.3                 729 MB/s  3616 MB/s     1227495  22.96 silesia/xml

      41681.143709      task-clock (msec)         #    0.991 CPUs utilized          
               505      context-switches          #    0.012 K/sec                  
                11      cpu-migrations            #    0.000 K/sec                  
            291159      page-faults               #    0.007 M/sec                  
      123153166169      cycles                    #    2.955 GHz                      (75.03%)
      275849355190      instructions              #    2.24  insn per cycle           (75.04%)
       49297484532      branches                  # 1182.729 M/sec                    (75.02%)
        1668515452      branch-misses             #    3.38% of all branches          (75.02%)

      42.073131221 seconds time elapsed


Adding this patch: 

diff --git a/lib/lz4.c b/lib/lz4.c
index ddaab77..ccc697e 100644
--- a/lib/lz4.c
+++ b/lib/lz4.c
@@ -1649,8 +1649,9 @@ LZ4_decompress_generic(
 
                 /* Fastpath check: Avoids a branch in LZ4_wildCopy32 if true */
                 if (!(dict == usingExtDict) || (match >= lowPrefix)) {
-                    if (offset >= 16) {
-                        memcpy(op, match, 16);
+                    if (offset >= 8) {
+                        memcpy(op, match, 8);
+                        memcpy(op+8, match+8, 8);
                         memcpy(op+16, match+16, 2);
                         op += length;
                         continue;

lz4 1.8.3                 329 MB/s  2991 MB/s     6428742  63.07 silesia/dickens
lz4 1.8.3                 536 MB/s  2867 MB/s    26435667  51.61 silesia/mozilla
lz4 1.8.3                 514 MB/s  3495 MB/s     5440937  54.57 silesia/mr
lz4 1.8.3                 901 MB/s  4321 MB/s     5533040  16.49 silesia/nci
lz4 1.8.3                 438 MB/s  2958 MB/s     4338918  70.53 silesia/ooffice
lz4 1.8.3                 469 MB/s  3150 MB/s     5256666  52.12 silesia/osdb
lz4 1.8.3                 333 MB/s  2574 MB/s     3181387  48.00 silesia/reymont
lz4 1.8.3                 568 MB/s  3571 MB/s     7716839  35.72 silesia/samba
lz4 1.8.3                 446 MB/s  3977 MB/s     6790273  93.63 silesia/sao
lz4 1.8.3                 499 MB/s  3067 MB/s   100881331  47.60 silesia/silesia.tar
lz4 1.8.3                 378 MB/s  2819 MB/s    20139988  48.58 silesia/webster
lz4 1.8.3                1234 MB/s  9311 MB/s     8390195  99.01 silesia/x-ray
lz4 1.8.3                 729 MB/s  3645 MB/s     1227495  22.96 silesia/xml


      41687.349560      task-clock (msec)         #    0.991 CPUs utilized          
               505      context-switches          #    0.012 K/sec                  
                 9      cpu-migrations            #    0.000 K/sec                  
            291158      page-faults               #    0.007 M/sec                  
      120422000483      cycles                    #    2.889 GHz                      (75.01%)
      274586415771      instructions              #    2.28  insn per cycle           (75.03%)
       47621821770      branches                  # 1142.357 M/sec                    (75.03%)
        1566369291      branch-misses             #    3.29% of all branches          (75.04%)

      42.076941335 seconds time elapsed
```

The overall benchmark numbers look neutral-ish to me, but indeed the total branches and branch misses drops a bit.  I can add the above patch.  Thanks!